Major Changes to the Project
