+++ Simulator/__init__.py
@@ -1,0 +1,1 @@
+from command_line_script import *
